backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.0
h11==0.16.0
idna==3.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
sqlalchemy==2.0.30
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
cryptography
openai
python-dotenv
python-multipart